// clients/typescript/generated/models/ChatCompletionRequestMessageFunctionCall.ts
/**
 * OpenAI API
 * APIs for sampling from and fine-tuning language models
 *
 * OpenAPI spec version: 2.0.0
 * Contact: blah+oapicf@cliffano.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* The name and arguments of a function that should be called, as generated by the model.
*/
export class ChatCompletionRequestMessageFunctionCall {
    /**
    * The name of the function to call.
    */
    'name'?: string;
    /**
    * The arguments to call the function with, as generated by the model in JSON format. Note that the model does not always generate valid JSON, and may hallucinate parameters not defined by your function schema. Validate the arguments in your code before calling your function.
    */
    'arguments'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "arguments",
            "baseName": "arguments",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ChatCompletionRequestMessageFunctionCall.attributeTypeMap;
    }

    public constructor() {
    }
}