// clients/typescript/generated/models/CreateChatCompletionRequestStop.ts
/**
 * OpenAI API
 * APIs for sampling from and fine-tuning language models
 *
 * OpenAPI spec version: 2.0.0
 * Contact: blah+oapicf@cliffano.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Up to 4 sequences where the API will stop generating further tokens. 
*/
export class CreateChatCompletionRequestStop {

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
    ];

    static getAttributeTypeMap() {
        return CreateChatCompletionRequestStop.attributeTypeMap;
    }

    public constructor() {
    }
}